Flatpak runtime 20.08

See merge request picviz/inendi-inspector!823

+++ libpvparallelview/tests/gl_timeseries.cpp
@@ -73,7 +73,7 @@ int main(int argc, char** argv)
 	                                  view->get_real_output_selection());
 
 	std::unordered_set<size_t> selected_timeseries;
-	for (int i = 0; i < timeseries.size(); ++i) {
+	for (size_t i = 0; i < timeseries.size(); ++i) {
 		selected_timeseries.emplace(i);
 	}
 	sampler.set_selected_timeseries(selected_timeseries);
